Refactored mute filters
