+++ backend/filters/__init__.py
@@ -1,4 +1,7 @@
 # __all__ = []
 
+from .filter import Filter
 from .api_test import FilterAPITestFilter
 from .edge_outline import EdgeOutlineFilter
+from .rotate import RotationFilter
+from .mute import MuteAudioFilter, MuteVideoFilter
