+++ backend/filters/filter_factory.py
@@ -6,7 +6,6 @@ if TYPE_CHECKING:
     from modules.track_handler import TrackHandler
 
 from filters import *
-from filters.mute import MuteAudioFilter, MuteVideoFilter
 
 from custom_types.filters import FilterDict
 from modules.exceptions import ErrorDictException
