+++ backend/filters/mute/__init__.py
@@ -1,0 +1,2 @@
+from .mute_audio_filter import MuteAudioFilter
+from .mute_video_filter import MuteVideoFilter
